Added main mini shell program
Parses the command line to remove extra whitespace and to check if a command should run in the background. If it is supposed to run in the background, it forks and parent continues to receive input. Pipes have been implemented so it is possible to pipe output from one command into the input of another.