Added program that generates scraping script
Generates a script that can be run by the MiniShell to find the number of unique words in all Wikipedia pages that have two-char long names (ie AA, AB, AC, ... , ZY, ZZ). It uses grep to match the Wikipedia page link to a regular expression, then uses wget to access the website. Finally, it also uses lynx to convert the .html file to a .txt file without the HTTP header. At which point the script then generates code to sort all the words by count, after finding the counts of unique words. The number of arrays I used for the commands came as a result of the max line size in the MiniShell program, since I could not have commands that were too long since they would cause Segmentation Faults.